requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.3.0
faiss-cpu>=1.7.4
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# PDF processing
PyMuPDF>=1.23.0
pdf2image>=1.16.3
Pillow>=10.0.0

# OCR support
pytesseract>=0.3.10
easyocr>=1.7.0

# Google Drive integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Utilities
numpy>=1.24.0
python-dateutil>=2.8.2
requests>=2.31.0
tqdm>=4.66.0

